3779. Cleanups for the contrib programs recently moved into the main xc/       tree (#3701, Bruno Haible). 3778. Small lynx.cf fix (#3700, Thomas Mueller). 3777. Initialise MMIO32 for r128 when using fbdev (#3699, Kostas Gewrgiou). 3776. Avoid most lockups for Trio3D startup (s3virge driver) (#3698,       Harald Koenig). 3775. Fix unresolved and conflicting symbols in X-TT, and fix the broken       BIG5 converter (#3697, Takuya Shiozaki). 3774. Update README.fonts (#3696, Juliusz Chroboczek). 3773. Add dot, line, arc accelerations for 24bpp in fb, and cleanup most       warnings (#3695, Keith Packard). 3772. Make it possible to build the r128 driver with support for fb       (disabled by default) (#3693, Keith Packard). 3771. Fix a problem with lnx_ev56 (#3693, Keith Packard). 3770. Fix references to noPanoramiXExtension when Xinerama is not built       (#3691, Christopher Sekiya). 3769. Fix possible races in xauth and libXau (#3690, 3694, Colin Phipps). 3768. Fix some Xv and PC-98 init bugs in the apm driver (#3689, Loic Grenie). 3767. Fix SVR4.0 build problem in the cyrix driver (David Dawes). 3766. Fix SVR4.0 build problems in the sis driver (#3687, Takaaki Nomura).

+++ programs/Xserver/fb/fbgetsp.c
@@ -1,5 +1,5 @@
 /*
- * $Id: fbgetsp.c,v 1.2 2000/01/21 01:11:57 dawes Exp $
+ * $Id: fbgetsp.c,v 1.3 2000/02/14 19:20:29 dawes Exp $
  *
  * Copyright © 1998 Keith Packard
  *
@@ -21,7 +21,7 @@
  * TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
  * PERFORMANCE OF THIS SOFTWARE.
  */
-/* $XFree86: xc/programs/Xserver/fb/fbgetsp.c,v 1.1 1999/11/19 13:53:43 hohndel Exp $ */
+/* $XFree86: xc/programs/Xserver/fb/fbgetsp.c,v 1.2 2000/01/21 01:11:57 dawes Exp $ */
 
 #include "fb.h"
 
@@ -49,7 +49,7 @@ fbGetSpans(DrawablePtr	pDrawable,
     
     while (nspans--)
     {
-	xoff = (((int) pchardstStart) & (FB_MASK >> 3));
+	xoff = (int) (((long) pchardstStart) & (FB_MASK >> 3));
 	dst = (FbBits *) (pchardstStart - xoff);
 	xoff <<= 3;
 	fbBlt (src + ppt->y * srcStride, srcStride,
